Updated Qty to decimal

+++ src/DelayedPaymentService/Data/ViewModels/PurchasedProduct/PurchasedProductViewModel.cs
@@ -4,7 +4,7 @@
     {
         public string Id { get; set; } = null!;
 
-        public int BoughtQuantity { get; set; }
+        public decimal BoughtQuantity { get; set; }
 
         public decimal PurchasePrice { get; set; }
